Add working etch-a-sketch
The etch-a-sketch now functions as expected.
The amount of squares can be modified in the code
to adjust the drawing space without changing the
size of the canvas.

Now I must add a button that sends a prompt, to
allow the user to change the amount of squares on
the canvas, and subsquently add a feature for
dynamic colors.