Add random RGB values to grid and darkening effect
When the user draws on the canvas, it will
generate a random color, applying a rainbow
effect.

Furthermore, each time the user brushes
over a previously colored pixel it will darken
the opacity by 10%, eventually achieving a fully
black square on the 11th brush.

Finally, I added some CSS styling to the overall
project to give it some life.

This marks the end of my Etch-a-Sketch project for
now. Really happy with the way it turned out!